requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
filelock==3.20.0
Flask==3.1.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.2.16
langchain-classic==1.0.1
langchain-community==0.2.16
langchain-core==0.3.83
langchain-groq==0.3.2
langchain-huggingface==0.0.3
langchain-pinecone==0.2.0
langchain-tests==0.3.22
langchain-text-splitters==0.2.4
langsmith==0.7.7
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib-inline==0.1.7
mdurl==0.1.2
-e git+https://github.com/Ughh-Sarthak/Medical-Chatbot.git@2f95f273e3aa7c5597b958ba138341c1a8b0d87b#egg=medical_chat_bot
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
openai==1.109.1
orjson==3.11.5
packaging==24.2
parso==0.8.4
pillow==12.1.0
pinecone==7.3.0
pinecone-client==5.0.1
pinecone-plugin-assistant==1.8.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.48
propcache==0.4.1
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.18.0
pypdf==5.6.1
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==308
PyYAML==6.0.3
pyzmq==26.2.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
stack-data==0.6.3
sympy==1.14.0
syrupy==4.9.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
vcrpy==7.0.0
wcwidth==0.2.13
Werkzeug==3.1.5
wheel==0.45.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.23.0
